backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
SQLAlchemy==2.0.36